feat: user testing (#8)
* fix: activities testing

* feat: add testing for users

* feat: improve UserManager methods and tests

* chore: improve user types class and tests

* fix: forget to add class User to the last commit

* fix: removeUser

* fix: date in test

+++ src/main/java/MakeItFit/users/UserInterface.java
@@ -20,7 +20,7 @@ public interface UserInterface {
 
         Gender getGender();
 
-        int getWeight();
+        float getWeight();
 
         int getHeight();
 
@@ -41,7 +41,7 @@ public interface UserInterface {
 
         void setGender(Gender gender);
 
-        void setWeight(int weight);
+        void setWeight(float weight);
 
         void setHeight(int height);
